federated_learning_sandbox/containers/data-prep/requirements.txt
pandas==1.3.4
numpy==1.21.3
sklearn==0.0
sklearn-pandas==1.8.0
scikit-learn==1.0.2
pydicom==2.2.0
Pillow==8.4.0